add interaction with state of js qraphql and put the data  into Charts
