+++ src/query/reactExperienceQuery.js
@@ -3,43 +3,43 @@ import {
   } from "@apollo/client";
 
 export const REACT_EXPERIENCE_QUERY = gql`
-  query react_experienceQuery { 
-      survey(survey: state_of_js) { 
-        tool(id: react) { 
-          id 
-          entity { 
-            homepage { 
-              name 
-              url 
-            } 
+query react_experienceQuery { 
+    survey(survey: state_of_js) { 
+      tool(id: react) { 
+        id 
+        entity { 
+          homepage { 
             name 
-            github { 
-              name 
-              description 
-              url 
-              homepage 
-            } 
+            url 
           } 
-          experience(filters: {country: {in: [PER, COL]}}) { 
-            all_years { 
-              year 
-              completion { 
+          name 
+          github { 
+            name 
+            description 
+            url 
+            homepage 
+          } 
+        } 
+        experience(filters: {country: {in: [PER, COL]}}) { 
+          all_years { 
+            year 
+            completion { 
+              count 
+            } 
+            facets { 
+              buckets { 
+                id 
                 count 
               } 
-              facets { 
-                buckets { 
-                  id 
-                  count 
-                } 
-                completion { 
-                  total 
-                  count 
-                  percentage_question 
-                } 
+              completion { 
+                total 
+                count 
+                percentage_question 
               } 
             } 
           } 
         } 
       } 
     } 
+  } 
 `;
